+++ CHANGES.md
@@ -15,6 +15,7 @@ Changes in 8.1
 - Add File Attached to Fees in StudentFees.php & Student_Billing/functions.inc.php
 - Fix User Widgets for Advanced Search / Print Class Lists exports in GetStaffList.fnc.php, GetStuList.fnc.php & Search.inc.php
 - Add Export fields list (form) & Export fields list + extra SQL (student list) action hooks in Export.php & Actions.php
+- Do not remove Full Day and Half Day school periods from the Schedule table in PrintSchedules.php
 
 Changes in 8.0.4
 ----------------